+++ src/components/Router/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 
-import Signup from '../../screens/Signup';
-import Login from '../../screens/Login';
+import SignupForm from '../../screens/SignupForm';
+import LoginForm from '../../screens/LoginForm';
 import Home from '../../screens/Home';
 import { PATHS } from '../../constants/paths';
 
@@ -10,8 +10,8 @@ function Router() {
   return (
     <BrowserRouter>
       <Switch>
-        <Route path={PATHS.login} component={Login} />
-        <Route path={PATHS.signup} component={Signup} />
+        <Route path={PATHS.login} component={LoginForm} />
+        <Route path={PATHS.signup} component={SignupForm} />
         <Route exact path={PATHS.root} component={Home} />
       </Switch>
     </BrowserRouter>
