--- src/screens/Login/index.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import LoginForm from '../../components/LoginForm';
-
-function Login() {
-  return <LoginForm />;
-}
-
-export default Login;
