+++ src/screens/LoginForm/i18n.js
