--- src/screens/Signup/index.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import SignupForm from '../../components/SignupForm';
-
-function Signup() {
-  return <SignupForm />;
-}
-
-export default Signup;
