+++ src/screens/SignupForm/styles.module.scss
@@ -3,9 +3,7 @@
 .signup-form {
   background-color: $white-main;
   border-top: 6px solid $blue;
-  display: flex;
-  flex-direction: column;
-  margin: 30px auto;
+  margin-top: 30px;
   max-width: 300px;
   padding: 40px 20px 10px;
 }
